add find-time-zone utility function, move find week-range to utils
