+++ package.json
@@ -10,6 +10,7 @@
     "stage": "node server.js"
   },
   "dependencies": {
+    "axios": "^1.7.8",
     "bcryptjs": "2.4.3",
     "body-parser": "1.20.3",
     "cors": "2.8.5",
